Add card content interaction (#261)
* Add blurry background to NewsCollectionViewCell

* Add interaction to various data sources

When an item of a data source is clicked, the data source receives a
callback and can determine how to proceed, e.g. by displaying the link
in a SafariViewController or by opening a ViewController that shows
additional details.

* Add background to calendar items

* Improve countdown formatting in calendar items

Now, we only show the days, hours and minutes until the calendar event.
We don’t display seconds in the countdown anymore.

* Add placeholder text for empty cafeteria menus

* Add "Show all" button in cards

The button opens a ViewController with details.

* Add "Show all" for cafeteria card

* Preload some ViewControllers with data

When opening some ViewControllers, for instance news or MVG stations,
from a card, we pass along the already downloaded information.

* Limit number of cells in card CollectionViews

The CollectionViews in cards now only display a limited number of cells.
For single items, the limit is 10. For composite items (items that
contain other items, such as MVG stations and cafeteria), the limit is
5.

* Added shadows to tu-film, news, newsspread and calendar

* fixed merge conflict

* Remove shadow on TUFilm posters

* Fix problem that prevented clicks on movie posters

* Add show all button for lectures

* Add show all button for grades

* Pass grades to detail view when tapping show more

* Add direct selection of study rooms

* Add direct selection of lectures

* Improve design of calendar cards

- Minimum font scale for autoshrink is now 0.8, not 0.5
- Use short weekday in date label ("Sat" instead of "Saturday")

* Remove testing code